formulaire multietapes

+++ database/migrations/2026_08_19_100701_create_abonnements_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->date('date_debut');
             $table->date('date_fin')->nullable();
             $table->enum('etat', ['ACTIF', 'INACTIF'])->default('INACTIF');
-            $table->unsignedBigInteger('menage_id');
+            $table->integer('menage_id');
 
             $table->foreign('menage_id')
                 ->references('id')
